logs geração de thumbnails e correções

+++ jobs/icon_generator/icon_generator.go
@@ -3,6 +3,7 @@ package icongenerator
 import (
 	"errors"
 	"fmt"
+	"log"
 	"mime"
 	"os"
 	"path/filepath"
@@ -18,28 +19,26 @@ import (
 )
 
 func RunGenerator() error {
-	erro := processFile(config.AbsRootDir, database.DB)
-	if erro != nil {
-		return erro
-	}
-	return nil
+	log.Print("starting files process")
+	return processFile(config.AbsRootDir, database.DB)
 }
 
 func processFile(path string, db *gorm.DB) error {
+	log.Printf("reading dir %s", path)
 	files, err := os.ReadDir(path)
 	if err != nil {
-		return err
+		return fmt.Errorf("error process file: %s, %s", err, path)
 	}
 
 	parentPath := path[len(config.AbsRootDir):]
 	filesDb, err := repository.GetFilesMapFromParentPath(db, parentPath)
-	dirs := make([]string, 0)
+	log.Printf("%d file(s) were found in the database.", len(filesDb))
+
 	for _, file := range files {
 		filePath := filepath.Join(parentPath, file.Name())
 		fullPath := filepath.Join(config.AbsRootDir, filePath)
 
 		if file.IsDir() {
-			dirs = append(dirs, filePath)
 			err := processFile(fullPath, db)
 
 			if err != nil {
@@ -51,27 +50,30 @@ func processFile(path string, db *gorm.DB) error {
 		fileEntity, exists := filesDb[file.Name()]
 		if !exists {
 			fileEntity = *models.NewFile(file.Name(), filePath, &parentPath)
+			log.Printf("file %s will be created in the database.", filePath)
 		} else {
 			delete(filesDb, file.Name())
+			log.Printf("file %s already exists.", filePath)
 		}
 
-		err := db.Save(&fileEntity).Error
-		if err != nil {
-			return err
+		if err := db.Save(&fileEntity).Error; err != nil {
+			return fmt.Errorf("error save file in db: %s (%s, %s)", err, file.Name(), path)
 		}
 
 		contentType := mime.TypeByExtension(filepath.Ext(file.Name()))
 		gen, ok := generator.GetGenerator(contentType)
 		if !ok {
+			log.Printf("no generator found to file %s.", contentType)
 			continue
 		}
 
 		ok, err = doGenerateIcon(&fileEntity, fullPath, db)
 		if err != nil {
-			return err
+			return fmt.Errorf("error save file doGenerateIcon: %s (%s, %s)", err, file.Name(), path)
 		}
 
 		if !ok {
+			log.Printf("file %s mark to no generate icon", fullPath)
 			continue
 		}
 		if fileEntity.IconPath == nil {
@@ -80,31 +82,34 @@ func processFile(path string, db *gorm.DB) error {
 		}
 
 		iconFullPath := filepath.Join(config.AbsRootDir, *fileEntity.IconPath)
+		log.Printf("generating icon to file %s saving in %s", fullPath, iconFullPath)
 		ok, err = generator.GenerateIcon(fullPath, iconFullPath, gen)
 
 		if err != nil {
-			return err
+			return fmt.Errorf("error save file generate icon: %s, (%s, %s)", err, fullPath, path)
 		}
 		if !ok {
+			log.Printf("icon %s was not generated", iconFullPath)
 			continue
 		}
 		fileEntity.SetIconPath(fileEntity.IconPath)
 
-		err = db.Save(&fileEntity).Error
-		if err != nil {
-			return err
+		if err = db.Save(&fileEntity).Error; err != nil {
+			return fmt.Errorf("error save file in db %s, %s", file.Name(), path)
 		}
 	}
 	for _, fileEntity := range filesDb {
-		if fileEntity.IconPath != nil {
+		if fileEntity.IconPath != nil && *fileEntity.IconPath != "" {
 			fileIconPath := filepath.Join(config.AbsRootDir, *fileEntity.IconPath)
+			log.Printf("delete icon from cache %s", fileIconPath)
 			err := os.Remove(fileIconPath)
 			if err != nil {
 				if !errors.Is(err, os.ErrNotExist) {
-					return err
+					return fmt.Errorf("error remove file %s", fileIconPath)
 				}
 			}
 		}
+		log.Printf("file deleted from database %s", fileEntity.FilePath)
 		db.Delete(fileEntity)
 	}
 	return nil
@@ -137,9 +142,11 @@ func doGenerateIcon(fileEntity *models.File, fullPath string, db *gorm.DB) (bool
 
 	fileEntity.CreatedAtUnix = createdAt
 	fileEntity.UpdatedAtUnix = updatedAt
-	fileEntity.SetGenerateIcon()
-	err = db.Save(&fileEntity).Error
-	if err != nil {
+	if err := fileEntity.SetGenerateIcon(); err != nil {
+		return false, err
+	}
+
+	if err = db.Save(&fileEntity).Error; err != nil {
 		return false, err
 	}
 	return true, nil
